+++ src/ios/SQLitePlugin.h
@@ -38,13 +38,18 @@ typedef int WebSQLError;
 -(void) close: (CDVInvokedUrlCommand*)command;
 -(void) delete: (CDVInvokedUrlCommand*)command;
 
+#if 0 // XXX NOT NEEDED:
 -(void) openNow: (CDVInvokedUrlCommand*)command;
 -(void) closeNow: (CDVInvokedUrlCommand*)command;
 -(void) deleteNow: (CDVInvokedUrlCommand*)command;
+#endif
 
 // Batch processing interface
 -(void) backgroundExecuteSqlBatch: (CDVInvokedUrlCommand*)command;
 
+#if 0 // XXX NOT NEEDED:
 -(void) executeSqlBatchNow: (CDVInvokedUrlCommand*)command;
+// ...
+#endif
 
 @end /* vim: set expandtab : */